Fix bug with sqlite getting experation. Finalize API and Tests

+++ src/MonkeyCache.FileStore/Barrel.cs
@@ -47,11 +47,8 @@ namespace MonkeyCache.FileStore
 		/// <param name="data">Data object to store</param>
 		/// <param name="expireIn">Time from UtcNow to expire entry in</param>
 		/// <param name="eTag">Optional eTag information</param>
-		public void Add(string key, string data, TimeSpan expireIn, string eTag = null)
+		void Add(string key, string data, TimeSpan expireIn, string eTag = null)
 		{
-			if (data == null)
-				return;
-
 			indexLocker.EnterWriteLock();
 
 			try
@@ -80,13 +77,29 @@ namespace MonkeyCache.FileStore
 		/// <param name="expireIn">Time from UtcNow to expire entry in</param>
 		/// <param name="eTag">Optional eTag information</param>
 		/// <param name="jsonSerializationSettings">Custom json serialization settings to use</param>
-		public void AddObject<T>(string key, 
+		public void Add<T>(string key, 
 							T data, 
 							TimeSpan expireIn, 
 							string eTag = null,
 							JsonSerializerSettings jsonSerializationSettings = null)
 		{
-			var dataJson = JsonConvert.SerializeObject(data, jsonSerializationSettings ?? jsonSettings);
+
+			if (string.IsNullOrWhiteSpace(key))
+				throw new ArgumentException("Key can not be null or empty.", nameof(key));
+
+			if (data == null)
+				throw new ArgumentNullException("Data can not be null.", nameof(data));
+
+			var dataJson = string.Empty;
+
+			if (Utils.IsString(data))
+			{
+				dataJson = data as string;
+			}
+			else
+			{
+				dataJson = JsonConvert.SerializeObject(data, jsonSerializationSettings ?? jsonSettings);
+			}
 
 			Add(key, dataJson, expireIn, eTag);
 		}
@@ -104,6 +117,9 @@ namespace MonkeyCache.FileStore
 			{
 				foreach (var k in key)
 				{
+					if (string.IsNullOrWhiteSpace(k))
+						continue;
+
 					File.Delete(Path.Combine(baseDirectory.Value, Hash(k)));
 					index.Remove(k);
 				}
@@ -181,6 +197,9 @@ namespace MonkeyCache.FileStore
 		/// <returns>If the key exists</returns>
 		public bool Exists(string key)
 		{
+			if (string.IsNullOrWhiteSpace(key))
+				throw new ArgumentException("Key can not be null or empty.", nameof(key));
+
 			var exists = false;
 
 			indexLocker.EnterReadLock();
@@ -197,32 +216,6 @@ namespace MonkeyCache.FileStore
 			return exists;
 		}
 
-		/// <summary>
-		/// Gets the string entry for the specified key.
-		/// </summary>
-		/// <param name="key">Unique identifier for the entry to get</param>
-		/// <returns>The string that was stored if found, else null</returns>
-		public string Get(string key)
-		{
-			string result = null;
-
-			indexLocker.EnterReadLock();
-
-			try
-			{
-				var hash = Hash(key);
-				var path = Path.Combine(baseDirectory.Value, hash);
-
-				if (index.ContainsKey(key) && File.Exists(path))
-					result = File.ReadAllText(path);
-			}
-			finally
-			{
-				indexLocker.ExitReadLock();
-			}
-
-			return result;
-		}
 
 		/// <summary>
 		/// Gets the data entry for the specified key.
@@ -230,8 +223,11 @@ namespace MonkeyCache.FileStore
 		/// <param name="key">Unique identifier for the entry to get</param>
 		/// <param name="jsonSerializationSettings">Custom json serialization settings to use</param>
 		/// <returns>The data object that was stored if found, else default(T)</returns>
-		public T GetObject<T>(string key, JsonSerializerSettings jsonSerializationSettings = null)
+		public T Get<T>(string key, JsonSerializerSettings jsonSerializationSettings = null)
 		{
+			if (string.IsNullOrWhiteSpace(key))
+				throw new ArgumentException("Key can not be null or empty.", nameof(key));
+
 			var result = default(T);
 
 			indexLocker.EnterReadLock();
@@ -244,6 +240,12 @@ namespace MonkeyCache.FileStore
 				if (index.ContainsKey(key) && File.Exists(path))
 				{
 					var contents = File.ReadAllText(path);
+					if (Utils.IsString(result))
+					{
+						object final = contents;
+						return (T)final;
+					}
+
 					result = JsonConvert.DeserializeObject<T>(contents, jsonSerializationSettings ?? jsonSettings);
 				}
 			}
@@ -262,8 +264,8 @@ namespace MonkeyCache.FileStore
 		/// <returns>The expiration date if the key is found, else null</returns>
 		public DateTime? GetExpiration(string key)
 		{
-			if (key == null)
-				return null;
+			if (string.IsNullOrWhiteSpace(key))
+				throw new ArgumentException("Key can not be null or empty.", nameof(key));
 
 			DateTime? date = null;
 
@@ -289,8 +291,8 @@ namespace MonkeyCache.FileStore
 		/// <returns>The ETag if the key is found, else null</returns>
 		public string GetETag(string key)
 		{
-			if (key == null)
-				return null;
+			if (string.IsNullOrWhiteSpace(key))
+				throw new ArgumentException("Key can not be null or empty.", nameof(key));
 
 			string etag = null;
 
@@ -316,6 +318,9 @@ namespace MonkeyCache.FileStore
 		/// <returns>If the expiration data has been met</returns>
 		public bool IsExpired(string key)
 		{
+			if (string.IsNullOrWhiteSpace(key))
+				throw new ArgumentException("Key can not be null or empty.", nameof(key));
+
 			var expired = true;
 
 			indexLocker.EnterReadLock();
